feat: Finaliza Pagina de Login e Cadastros sem Back-end

+++ src/Pages/cadastros.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import Swal from 'sweetalert2';
 
 function cadastros({ history }) {
   const [email, setEmail] = useState('');
@@ -13,16 +14,30 @@ function cadastros({ history }) {
     setPassword(target.value);
   };
 
+  const usuarios = JSON.parse(localStorage.getItem('user'));
+
   const formValidation = () => {
     const validEmail = email.includes('@') && email.includes('.com');
     const seven = 7;
     return !!(!validEmail || password.length < seven);
   };
-
   const handleClick = () => {
-    localStorage.setItem('user', JSON.stringify({ email, password }));
-    history.push('/home');
+    if (usuarios === null) {
+      localStorage.setItem('user', JSON.stringify({ email, password }));
+      history.push('/home');
+    } else if (usuarios.email === email) {
+      Swal.fire({
+        title: 'Error!',
+        text: 'Email já Cadastrado',
+        icon: 'error',
+        confirmButtonText: 'Ok',
+      });
+    } else {
+      localStorage.setItem('user', JSON.stringify({ email, password }));
+      history.push('/home');
+    }
   };
+
   return (
     <section>
       <label htmlFor="inputsLogin">
